+++ test.py
@@ -144,7 +144,7 @@ pwsh -NoProfile -NoLogo -File /tmp/version-test.ps1''' % args.verify_version))
 
             if args.docker:
                 docker_run(distro_details['container_image'], '/omi/%s' % os.path.basename(temp_fd.name),
-                    env={'KRB5_CONFIG': '/omi/krb5.conf'}, interactive=args.interactive)
+                    env={'KRB5_CONFIG': '/omi/krb5.conf'}, interactive=args.interactive, shell=distro_details['shell'])
 
             else:
                 print("Running tests locally")