+++ CHANGELOG.md
@@ -14,6 +14,7 @@ It documents the changes in each of the tagged releases
 
 + Created `PSWSMan` which is a PowerShell module uploaded to the PowerShell Gallery that can install and manage the OMI libraries for you
 + Build `libpsrpclient` as well and add it to the release artifacts
++ Added Alpine 3 to the build matrix
 
 ## 1.2.1 - 2020-09-26